Fix #4709 and #3253: report CyclicError in implicit search better
- Record in CyclicReference if it's emitted during implicit search.
- Before checking if the error is on an implicit, check if we have an error
  during implicit search that happens while inferring a return type; that can
  happen whether the member we're typechecking is implicit or not, and needs the
  same cure.
- Add ErrorMessagesTests as requested

This improves the error message, and also puts more effort into diagnosing the
cycle: we now use similar logic as for recursive/overloaded members to check
we're indeed dealing with a (potentially) recursive implicit.

Remaining `CyclicReferenceInvolvingImplicit` are unexpected and might involve
bugs, so make the error message and explanation more tentative, following
`CyclicReferenceInvolving`.