Remove `NonEntity` in Dottydoc
This special entity was used to mean that an entity needs to be removed
from the doc AST after the transformation. Instead, we refactor the
DocMiniPhases to have them return a `List[Entity]` (rather than a single
`Entity`, as before).

It follows that removing an entity from the AST means returning an empty
list after transformation.

This change will be necessary to be able to support multiple `@usecase`
sections in the documentation, because it means that the transformation
has to introduce multiple `Entity`-es.